Move the main run as part of the orchestrator
