+++ lib/orchestrator.js
@@ -2,6 +2,7 @@
 
 var exec = require('child_process').exec;
 var Fs = require('fs');
+var Path = require('path');
 var Promise = require('promise');
 
 var DocumentDownloader = require('./document-downloader');
@@ -11,7 +12,24 @@ var SpecberusWrapper = require('../functions.js').SpecberusWrapper;
 var ThirdPartyChecker = require('../functions.js').ThirdPartyChecker;
 var TokenChecker = require('../functions.js').TokenChecker;
 
+var STATUS_ERROR = 'error';
 var STATUS_SUCCESS = 'success';
+
+// Configuration file
+require('../config.js');
+
+var argTempLocation = process.argv[2] || global.DEFAULT_TEMP_LOCATION;
+var argHttpLocation  = process.argv[3] || global.DEFAULT_HTTP_LOCATION;
+var argResultLocation = process.argv[5] || global.DEFAULT_RESULT_LOCATION;
+
+function Job() {
+  if (typeof this !== 'object') {
+    throw new TypeError('Job must be constructed via new');
+  }
+
+  this.status = '';
+  this.errors = [];
+}
 
 function dumpJobResult(dest, result) {
   Fs.writeFile(dest, JSON.stringify(result, null, 2) + '\n', function (err) {
@@ -271,6 +289,58 @@ Orchestrator.prototype.finishTasks = function (thisVersion, resultLocation) {
 
     return state;
   };
+};
+
+Orchestrator.prototype.run = function (spec, token) {
+  var self = this;
+
+  spec.jobs['retrieve-resources'] = new Job();
+  spec.jobs['specberus'] = new Job();
+  spec.jobs['token-checker'] = new Job();
+  spec.jobs['third-party-checker'] = new Job();
+  spec.jobs['publish'] = new Job();
+  spec.jobs['tr-install'] = new Job();
+  spec.jobs['update-tr-shortlink'] = new Job();
+
+  var tempLocation = argTempLocation + Path.sep + spec.id + Path.sep;
+  var resultLocation = argResultLocation + Path.sep + spec.id + '.json';
+  var httpLocation = argHttpLocation + '/' + spec.id + '/Overview.html';
+
+  return self.runDocumentDownloader(spec.url, tempLocation)(spec)
+    .then(self.runSpecberus(httpLocation))
+    .then(function (state) {
+      var metadata = state.metadata;
+
+      return self.runTokenChecker(
+        metadata.get('latestVersion'),
+        spec.url,
+        token
+      )(state)
+        .then(self.runThirdPartyChecker(httpLocation))
+        .then(self.runPublisher(metadata))
+        .then(self.runTrInstaller(metadata.get('thisVersion'), tempLocation))
+        .then(self.runShortlink(metadata.get('thisVersion')))
+        .then(self.finishTasks(metadata.get('thisVersion'), resultLocation));
+    })
+    .catch(function (err) {
+      console.log(err.stack);
+
+      spec.status = STATUS_ERROR;
+      var cmd =
+        global.SENDMAIL + ' ERROR ' + global.MAILING_LIST + ' ' + spec.url +
+        ' \'' + JSON.stringify(spec, null, 2) + '\'';
+
+      exec(cmd, function (err, stdout, stderr) {
+        if (err) console.error(stderr);
+      });
+
+      spec.history = spec.history.add(
+        'A system error occurred during the process.'
+      );
+
+      dumpJobResult(resultLocation, spec);
+      return Promise.reject(new Error('Orchestrator has failed.'));
+    });
 };
 
 module.exports = Orchestrator;
